// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsDeviceModeSchedulesItems_Dtos_PacsDeviceModeSchedulesItemEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Pacs_PacsDeviceModeSchedulesItems_Dtos_PacsDeviceModeSchedulesItemEditDto = {
    id?: string | null;
    pacsActionPlanGroupId?: string;
    deviceId?: string;
    deviceTypeId?: string;
    deviceTypeFunctionId?: string;
};